% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rank.r
\name{get_rank}
\alias{get_rank}
\alias{get_rank.default}
\alias{get_rank.zonation}
\title{Extract the rank raster}
\usage{
get_rank(x)

\method{get_rank}{default}(x)

\method{get_rank}{zonation}(x)
}
\arguments{
\item{x}{a zonation object.}
}
\description{
Extract the rank raster from a zonation object
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: extract the rank raster

\item \code{zonation}: extract the rank raster
}}
\examples{
library(raster)
r1 <- raster(matrix(runif(200^2, 0, 1), 200))
r2 <- raster(matrix(runif(200^2, 0, 1), 200))
plan <- zonation(stack(r1, r2))
plan_rank <- get_rank(plan)
plan_rank

}

