% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.zonation.r
\name{mean.zonation}
\alias{mean.zonation}
\title{Calculate mean proportion remaining}
\usage{
\method{mean}{zonation}(x, p = 0.1, features = NULL,
  method = c("arithmetic", "geometric"), ...)
}
\arguments{
\item{x}{a zonation object.}

\item{p}{numeric proportion of landscape protected.}

\item{features}{numeric vector of features id number (order features are added to rzonation)}

\item{method}{the method of calculating the mean arithmetic or geometric.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Calculate the mean proportion of features remaining when a given proportion of landscape is protected.
}
\examples{
library(raster)
r1 <- raster(matrix(runif(200^2, 0, 1), 200))
r2 <- raster(matrix(runif(200^2, 0, 1), 200))
plan <- zonation(stack(r1, r2))
mean(plan)

}

