% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.zonation.r
\name{plot.zonation}
\alias{plot.zonation}
\title{Plot a zonation object}
\usage{
\method{plot}{zonation}(obj, features = NULL, feature.names = NULL,
  invert = TRUE, main = "Performance curves",
  legend.title = "Performance", blackwhite = FALSE, ...)
}
\arguments{
\item{obj}{a zonation object.}

\item{features}{a numeric vector of columns to call from curves data.frame.}

\item{feature.names}{character vector of feature names to select for plotting}

\item{invert}{logical invert x-axis.}

\item{main}{title of plot, default is 'Performance curves'.}

\item{legend.title}{title of legend, default is 'Performance'.}

\item{blackwhite}{FALSE make plot black and white.}

\item{...}{other plot arguments.}
}
\description{
Plot a zonation object using ggplot
}
\examples{
library(raster)
r1 <- raster(matrix(runif(200^2, 0, 1), 200))
r2 <- raster(matrix(runif(200^2, 0, 1), 200))
plan <- rzonation::zonation(stack(r1, r2))
plot(plan)
}

