% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_curves.r
\name{plot_curves}
\alias{plot_curves}
\title{Plot a zonation object}
\usage{
plot_curves(
  x,
  features = NULL,
  feature_names = NULL,
  invert = TRUE,
  main = "Performance curves",
  legend.title = "Performance",
  blackwhite = FALSE,
  ...
)
}
\arguments{
\item{x}{a zonation object.}

\item{features}{a numeric vector of columns to call from curves data.frame.}

\item{feature_names}{character vector of feature names to select for plotting}

\item{invert}{logical invert x-axis.}

\item{main}{title of plot, default is 'Performance curves'.}

\item{legend.title}{title of legend, default is 'Performance'.}

\item{blackwhite}{FALSE make plot black and white.}

\item{...}{other plot arguments.}
}
\description{
Plot a zonation object using ggplot
}
\examples{
library(raster)
r1 <- raster(matrix(runif(200^2, 0, 1), 200))
r2 <- raster(matrix(runif(200^2, 0, 1), 200))
plan <- rzonation::zonation(stack(r1, r2))
plot_curves(plan)

}
