% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.zonation.r
\name{summary.zonation}
\alias{summary.zonation}
\title{See output of zonation run}
\usage{
\method{summary}{zonation}(object, ...)
}
\arguments{
\item{object}{a zonation object.}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
See the run information produced by zonation
}
\examples{
library(raster)
r1 <- raster(matrix(runif(200^2, 0, 1), 200))
r2 <- raster(matrix(runif(200^2, 0, 1), 200))
plan <- zonation(stack(r1, r2))
library(ggplot2)
summary(plan)

}
