% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.zonation.r
\name{plot.zonation}
\alias{plot.zonation}
\title{Plot a zonation object}
\usage{
\method{plot}{zonation}(obj, features = NULL, feature.names = NULL,
  main = "Performance curves", legend.title = "Performance",
  blackwhite = FALSE, ...)
}
\arguments{
\item{obj}{a zonation object.}

\item{features}{a numeric vector of columns to call from curves data.frame.}

\item{blackwhite}{FALSE make plot black and white.}

\item{...}{other arguments.}

\item{features}{a character vector of names that correlate to biodiversity feature names (this is to make pretty plots).}
}
\description{
Plot a zonation object using ggplot
}
\examples{
library(raster)
r1 <- raster(matrix(runif(200^2, 0, 1), 200))
r2 <- raster(matrix(runif(200^2, 0, 1), 200))
plan <- rzonation::zonation(stack(r1, r2))
plot(plan)

}

