% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zonation.r
\name{zonation}
\alias{zonation}
\title{Run the program zonation from R}
\usage{
zonation(features)
}
\arguments{
\item{features}{file paths of the GeoTiff files for features in the conservation plan.}
}
\description{
Runs the zonation software for conservation planning on a set of GeoTiff files
and returns the output as a list.
}

