% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zonation.r
\docType{methods}
\name{zonation}
\alias{zonation}
\alias{zonation,RasterStack-method}
\alias{zonation,character-method}
\title{Run the program zonation from R}
\usage{
zonation(features)

\S4method{zonation}{RasterStack}(features)

\S4method{zonation}{character}(features)
}
\arguments{
\item{features}{RasterStack or file paths of the raster files for features in the conservation plan.}
}
\description{
Runs the zonation software for conservation planning on a set of GeoTiff files
and returns the output as a list.
}
\section{Methods (by class)}{
\itemize{
\item \code{RasterStack}: run the program zonation for a RasterStack

\item \code{character}: run the program zonation for raster files
}}
\examples{
library(raster)
r1 <- raster(matrix(rnorm(100, 10)))
r2 <- raster(matrix(rnorm(100, 10)))
plan <- zonation(stack(r1, r2))

}

