% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zonation.r
\docType{methods}
\name{zonation}
\alias{zonation}
\alias{zonation,RasterBrick-method}
\alias{zonation,RasterStack-method}
\alias{zonation,character-method}
\title{Run the program zonation from R}
\usage{
zonation(features, params = NULL, settings = NULL, dir = NULL,
  alpha = 0, dist_smooth = FALSE, kernel_width_mult = 1,
  command_args = NULL, additional_settings = NULL)

\S4method{zonation}{RasterStack}(features, params = NULL, settings = NULL,
  dir = NULL, alpha = 0, dist_smooth = FALSE, kernel_width_mult = 1,
  command_args = NULL, additional_settings = NULL)

\S4method{zonation}{RasterBrick}(features, params = NULL, settings = NULL,
  dir = NULL, alpha = 0, dist_smooth = FALSE, kernel_width_mult = 1,
  command_args = NULL, additional_settings = NULL)

\S4method{zonation}{character}(features, params = NULL, settings = NULL,
  dir = NULL, alpha = 0, dist_smooth = FALSE, kernel_width_mult = 1,
  command_args = NULL, additional_settings = NULL)
}
\arguments{
\item{features}{RasterStack or file paths of the raster files for features in the conservation plan.}

\item{params}{a matrix of feature parameter values. Matrix must have 5 columns and one row for each feature. If unset all values default to 1.See Zonation manual for details.}

\item{settings}{a named list of settings equivalent to the zonation settings file (see zonation manual for details and list of settings)}

\item{dir}{a directory to house tmp files.}

\item{alpha}{numeric uncertainty parameter.}

\item{dist_smooth}{logical. should distribution smoothing be used.}

\item{kernel_width_mult}{numeric. factor to multiply feature dispersal kernel widths by.}

\item{command_args}{character string of command line arguments. See zonation manual for details.}

\item{additional_settings}{additional settings.}
}
\description{
Run the zonation software for conservation planning on a set of files or raster objects
and return the output as a list.
}
\section{Methods (by class)}{
\itemize{
\item \code{RasterStack}: run the program zonation for a RasterStack

\item \code{RasterBrick}: run the program zonation for a RasterStack

\item \code{character}: run the program zonation for raster files
}}
\examples{
library(raster)
r1 <- raster(matrix(runif(200^2, 0, 1), 200))
r2 <- raster(matrix(runif(200^2, 0, 1), 200))
plan <- zonation(stack(r1, r2))

}

